GUI: I had to add to new strings, to make the applet usuable. Now the translators will kill me, but otherwise the users would have done so.
Fixed a bug that caused the MultiMeters to currupt the grid layout.
Fixed a potention segfault in the SensorAgent.
Applet configuration now gets automatically saved and restored.
Size of applet can be adjusted in preference dialog.
Hopefully my last commit before 2.1. There still is a problem with reading
disk performance in Linux Kernel 2.4, but this is probably too complicated
to fix before the release.

svn path=/trunk/kdebase/ksysguard/; revision=82597