BUG: 153512
It appears this code was a performance optimization that relied on a bug in qt that has since been fixed.
see http://archives.devshed.com/forums/kde-96/saving-overheads-when-resizing-svg-2318239.html

It didn't save much computation anyway -- all we do is change the value of one variable.

The code I'm removing waited for the second of two resize events. We no longer get two resize events, so we ended up completely ignoring the resize event.

svn path=/trunk/KDE/kdebase/workspace/; revision=746778