BUG: 201525
Fix a bug when there was a sensor that was in a sgrd file but no longer available in ksysguardd we added a sensor with an empty name list and then crash when trying to access this list.  This is actually the right file for this fix, the commit of BasicSensor.cc was a mistake as there was no diff.

svn path=/trunk/KDE/kdebase/workspace/; revision=1003832