Fix compilation with strict libc (such as musl).
Many of the functions we use implicitly are not part of the C standard
library, so we have to specify that we're pulling from POSIX, X/Open,
BSD (or whatever standard we're using) before we can rely on the
definition being available.

There's plenty of room for refactoring here since I mostly stopped at
just getting the functions being called to show up from their includes,
although I did replace rindex with strrchr since the latter is
standardized.

Differential Revision: https://phabricator.kde.org/D6631