Properly check if mdraid array is active.
Don't compare a bool with an int.  This causes this part of the
comparison to always evaluate to 'true'.

In this case the corrected check actually appears to be necessary
(inactive raid arrays won't show the "raid%d" type of output that is
meant to be copied into ->Level).  Also noted by Coverity, CID 256387.