softraid: Remove dead code and associated warning.
At some point the `in_devs` code may have controlled more than it
currently does, but right now the only logic for `in_devs` is associated
with either if/then branches that do nothing but update `in_devs`, or do
nothing at all.

I looked up the Coverity CID entry 3228 that the code references... it
turns out it was Coverity CID 253228, and was a warning that this code
was dead.  Oops.  At this point the only thing that it looks like it may
have been doing on purpose was to think about warning if an md device
array was used as a subcomponent of a RAID array, but there's currently
no code that does this.

Given that the code is dead, it's best to get rid of it and the
associated warnings.

Differential Revision: https://phabricator.kde.org/D6658