Try to read CPU clock from cpufreq/scaling_cur_freq instead of /proc/cpuinfo
Summary:
From kernel 4.13 on, /proc/cpuinfo always reports the static normative
CPU clock and not the current one. So try to read the frequency from
cpufreq and only fall back to cpuinfo it that fails.

BUG: 382561

Test Plan:
Started ksysguard, frequency now changes. If the file is not
readable, the frequency is read from cpuinfo as expected.

Reviewers: #plasma, sebas

Reviewed By: #plasma, sebas

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8153