Correction with the & problem in tabs
Summary:
BUG: 382512

KSysGuard does not render the ampersand character in tab titles

Test Plan:
Tested on Arch Linux with Plasma 5.11.5 and Qt 5.10

I added a tab with an & in the title and I checked that the & displays correctly.
I added a tab with two and more & and I checked that all the & displays correctly.
Also works when open a saved tab.

Reviewers: mlaurent, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10151