Make some strings easier to translate
Summary:
Some strings in ksysguard use this advanced feature of being actually
multiple strings separated by "\xc2\x9c" and then letting the UI use one
or another depending how much space is available.

This is a cool feature, but until this patch it was made in a way the
translatable strings actually contained "\xc2\x9c" that makes it
hard to translate correctly since it's not a trivial character to input.

With this patch I have splitted the character out of the translatable
strings

Test Plan: Works

Reviewers: #plasma, #localization, mart

Reviewed By: #plasma, mart

Subscribers: mart, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D12713