Add plugin showing nvidia GPU load stats
This plugin shows the process GPU load and GPU memory usage for users of
Nvidia graphic cards.

Data is fetched by running the (fortunately lightweight) process
nvidia-smi to extract data. If this binary is not available this plugin
does nothing and the columns will not be available.

It's not feasible to try try and reverse engineer nvidia-smi to recreate
the functionality to do it without the extra executable. The headers are
hidden within a giant proprietary SDK, and internally it's just opening
a completely arbitrary address in /dev/mem