Update value (of application-used-memory) before using it.
The *used* and *total* values are computed before the answer
has been processed; then those values are used for writing
the strings to the UI, so they always reflect the **previous**
value of mUsedApplication. This is slightly suprising (only when
reading the code, though, since you're unlikely to notice these
memory variations from watching the UI).

Move the update-from-answer code up a bit.