Delete Clients in KsysGuardDaemon destructor
Clients and plugins are both children of the daemon. The Client destructor
dereferences SensorProperties to unsubscribe from them. Because usually
the object tree looks something like plugin -> sensorcontainer -> sensorobject -> sensorproperty
we can not rely on the properties still being around when the clients are deleted.
This fixes the segfault in the unit test detected by the address sanitizer.