Also expose two-part sensors as three-part sensors in ksysguardd
ksystemstats expects sensors to consist of at least three parts:
subsystem, object, property. Sensors like "system/processors" break this
assumption, causing them to not be properly readable. Rather than trying
to come up with an elaborate way to handle this in ksystemstats, simply
expose them also as a three-part sensor so the mapping doesn't need to
change. These sensors are not very dynamic anyway so duplicating them
does not do much harm. The two-part version is kept around for backward
compatibility with the ksysguard UI.