Divide network speeds by 2
We update the network speed every 2s to match the daemon update rate.
However, this means that when calculating the difference between the
current and the previous, we get twice the amount since we are a
calculating the difference over 2s. So divide by 2 to get the byte rate
per second.


(cherry picked from commit cf19c92b6edf2d5f37dda667ec18f4dfecb13878)