Do not load multiple versions of the same plugin
In instantiatePlugins the plugins are ordered according to the configured
libraries paths of the environment. When we load the plugins we are currently
doing the opposite of what we want. We load first the plugin from the directory
with the highest priority, some time later we load the same plugin from a
folder with a lower priority and then use the containers from this plugin and throw
the existing containers away!