Include StorageVolumes by default
We are searching for a StorageDrive to see if it is on a HardDisk or or some
other type of device. However on some setups a Volume will not have StorageDrive
as an ancestor, this is for example the case when using lvm. So let's include
every volume by default and exclude only those for which we find a Drive that is
not a hard disk.


(cherry picked from commit 8747897a6d3bbc5e9be17ac542d1bdc3a25eaa34)