Force device statistics refresh rate to be always one second
Unfortunately the statistics refresh rate of a device is a global property. So
when it is changed from the outside we just force it back to one second. The
initial rate is saved and updated when something tries to change it in order
to restore it upon destruction.
Because NetworkmanagerQt just forwards the DBus signal we have to guard the
slot with a boolean to we do not trigger it ourselves.
BUG:433063
FIXED-IN:5.21.1