Add new lmsensors plugin
The names are still not good because that is what we get from libsensors.
But it now also handles minimium and maximum if possible.
To not show the same sensors that we already handle in other plugins,
I use a simple blacklist for now for the ones we want to exclude.