Add uptime sensor to system information plugin
The old update method is renamed to the more accurate "init". Updating of the uptime
happens in a new update() method. Because it's only a few lines I just #ifdef'ed them.
The uptime subsystem from ksysguardd can be removed, aside from the uptime it contained
counts of CPUs and Cores.